requirements.txt
Pyrogram~=1.2.9
python-kucoin~=2.1.3